Brought over functionality from original site
Brought over alot of the functionality that had apready been written.
All tests are passing however the admin views still need to be setup.

In addition, the 'acts_as_markup' and 'carrierwave' references have been
commented out of the Content model for the moment as the methods were
not being added to ActiveRecord::Base correctly